--- betterstyles-ts/certs/ca.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDtzCCAp+gAwIBAgIJAM2mE5TOxoFEMA0GCSqGSIb3DQEBCwUAMHIxCzAJBgNV
-BAYTAlVTMQswCQYDVQQIDAJXQTEQMA4GA1UEBwwHUmVkbW9uZDEPMA0GA1UECgwG
-T2ZmaWNlMRwwGgYDVQQLDBNPZmZpY2VFeHRlbnNpYmlsaXR5MRUwEwYDVQQDDAxs
-b2NhbGhvc3QtY2EwHhcNMTgwNDExMTgwNjExWhcNMjMwNDExMTgwNjExWjByMQsw
-CQYDVQQGEwJVUzELMAkGA1UECAwCV0ExEDAOBgNVBAcMB1JlZG1vbmQxDzANBgNV
-BAoMBk9mZmljZTEcMBoGA1UECwwTT2ZmaWNlRXh0ZW5zaWJpbGl0eTEVMBMGA1UE
-AwwMbG9jYWxob3N0LWNhMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA
-ymd3VtdKsAtRo2Ug5cpC2sXa2VACIbwkQHNEps2Q8wcBhAdfahKIStfQKrTF8UqW
-tcpfSIhxcBMBhBITDXKlnu1HRO77ufC/w1AToJKPfzC2RfMEiLpeuHk8Sy65ST8f
-o3L1BuqwYsVspFu//hLpeGQAv7Z4QV/DOeR2Uw8bi9m8ayYEDaMckVq1Pq7gF+vP
-tOG/EDyAHoM5PzWd75TNNM3ysbN0edT9SEk3fOUOR1b0M0QXsxRAzgkpFUtTFy4a
-7YMFL113Jb+wMiUS7KZJalDrcjhthJInlAUwYNE3qFGqK412VbGDnwnas7L3+m8u
-PXM6AMgFKRvT+48zf9JkyQIDAQABo1AwTjAdBgNVHQ4EFgQUAnM7hKUXgNu7/kXI
-TkhWbbx428IwHwYDVR0jBBgwFoAUAnM7hKUXgNu7/kXITkhWbbx428IwDAYDVR0T
-BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEAPcsXB8fcFuRf3OuX7iUFPU1c04VK
-1y0HnWNC0jkPQKAale5+AsV99TFIBxb4ml+iokUU+SIdMwlHFkeVeVHhH8545LUF
-lbRQCjlcWsN9eAAci7h+BxE9sQt4LBNGmedzM+7+9rR3pzMI6zsaVyK7xS85G1mF
-nw3Ajfs+zqu3li+wsVF/4M7o6z0mjA036ejh3YgLabMSNHfNA+1ltPZwlKSE6+zR
-Wudewkb2yeDK85Z4l3vOhZPtZsBV2Uni34qXdTSAPGSXcRKd9NuArTjvt1QbD313
-8YY1RxQjEDdW14StquAbm8BCrGZ6ks3h5mJQh++6pqELRI3P24e5zc1gIg==
------END CERTIFICATE-----
