--- betterstyles-ts/certs/server.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDljCCAn6gAwIBAgIJAIReiYJV7pQTMA0GCSqGSIb3DQEBCwUAMHIxCzAJBgNV
-BAYTAlVTMQswCQYDVQQIDAJXQTEQMA4GA1UEBwwHUmVkbW9uZDEPMA0GA1UECgwG
-T2ZmaWNlMRwwGgYDVQQLDBNPZmZpY2VFeHRlbnNpYmlsaXR5MRUwEwYDVQQDDAxs
-b2NhbGhvc3QtY2EwIBcNMTgwNDExMTgwNjQxWhgPMjExODAzMTgxODA2NDFaMG8x
-CzAJBgNVBAYTAlVTMQswCQYDVQQIDAJXQTEQMA4GA1UEBwwHUmVkbW9uZDEPMA0G
-A1UECgwGT2ZmaWNlMRwwGgYDVQQLDBNPZmZpY2VFeHRlbnNpYmlsaXR5MRIwEAYD
-VQQDDAlsb2NhbGhvc3QwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCz
-h4LApOQXf9U5xukxNpWjhn653YhcXjJ1asK1RTQwvYQoJWpFvrX12Z7EzydpkR8l
-TRGZmghfF6LtKlci8GNwActPC5v/0hO6IpabDdVntaEVrwO8VACkBQi1N1CrtL05
-6gz4Aj2xMLYp6k2DVz1UWxF40YnPsWSRHzW8OmtP0U5XLIiwrxh9G4lwuFFan+Ha
-346Hyjagt8ptc8S3FSRF/06J4iR9kvM6iCy9DWP44HPxQQFQ2VRN8E+TpJdevtx3
-F1JZ8GsawKKRbeX2o++j1GMGCkeUGe9G6YBX0mD3d/Bx69gm1E2SRmdiJzEWQoZC
-yzIAK72YF+QiHi1qdzSVAgMBAAGjMDAuMAkGA1UdEwQCMAAwCwYDVR0PBAQDAgXg
-MBQGA1UdEQQNMAuCCWxvY2FsaG9zdDANBgkqhkiG9w0BAQsFAAOCAQEAekvobOVm
-hFgd6io16OeXahtKDaCm1NMw1YJoFE5LmBZRqJf8gAHT58Vg470lyURFxcn0MxXO
-Ho/wQlqZ/ycele9zkp3GPl03+7gA9py5EPvYyfuElQB0rG6TwIsAOD76Fubzom+X
-t+NE8zwF2Py2iMy+4YbiJqck+GfLOVPRaXgJt1BvY70G7GYOxOu+fLJuEsyvbwfN
-OhyDxGdQL/1CqasV3AOb+60qmHAg/eDx+T3M4ruTvrLsrtoRHsONyg3NQOEvgOEO
-8s/7BdMwwa0l4swlF0ozTeyaq0wfEHa1N1Wh8xH3EKSsiyOZclsPe2FG348W6GQA
-Hmv1P1mWNdKYWg==
------END CERTIFICATE-----
