--- betterstyles-ts/certs/server.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCzh4LApOQXf9U5
-xukxNpWjhn653YhcXjJ1asK1RTQwvYQoJWpFvrX12Z7EzydpkR8lTRGZmghfF6Lt
-Klci8GNwActPC5v/0hO6IpabDdVntaEVrwO8VACkBQi1N1CrtL056gz4Aj2xMLYp
-6k2DVz1UWxF40YnPsWSRHzW8OmtP0U5XLIiwrxh9G4lwuFFan+Ha346Hyjagt8pt
-c8S3FSRF/06J4iR9kvM6iCy9DWP44HPxQQFQ2VRN8E+TpJdevtx3F1JZ8GsawKKR
-beX2o++j1GMGCkeUGe9G6YBX0mD3d/Bx69gm1E2SRmdiJzEWQoZCyzIAK72YF+Qi
-Hi1qdzSVAgMBAAECggEABNGLSd5jHNpaf719/ssnQwdWGWj/wvLoScx+w1zDJNfB
-V06BxrNysNGReFjTBvtVKawi38uJk/vPbKwpQI/RK21w5oM6QaiCzQkmn6zEYI84
-bpuO7xwWFFDqRerhKGqmdYp5pKJyw3G6Qe+++tp5k+0J0ANXvpJVT6vVxmKPBMpe
-3KkylxBqe7rNOorjYpfuLu4/z0jtz+xCOdI3y9F3Jl/FMyE6NAGCVEWKnoAK4EOv
-nz33jo8DUan+pSi/ssJbOqxfnkBE9yFcrytpMjTeCff/LcbHDi1gdhGn7iKc/QxE
-PA98yzSOZeWgR5RbBEJEQb5831VRr6Gt2b0qU9JldQKBgQDYxAbAkASMSViP5THp
-djk0+daExKQ4GVKMi5RI6BHQKJq75H7JWQ5zArvxL1AyEpMac0Q6w0FIxnbkiYKV
-tmc6kjmjbsgA+ml02nJtmfNizIf9Z9Op1dWfc3maRkiYqEl5GS0h7a2a/rtV+OC5
-olCwiLFy7ch0Ik/fjhTsy0fIBwKBgQDUBh0DUrrp6C+yeQZX9WIM9C7qZODZlYCU
-waUIibL/JPjv8zA5WBRjbvmn+QyafwjRbN+yiboeWbrdaNH7U1cyjb85O81G2SAM
-BK9yTiBuvRi1sk3H9KKG628Gjb93uyIN7htLcZGUYlkff5h2fYJHPpOd6lZU0bUs
-ZQZ6QaUfgwKBgCJTWGZh5/Z3HukaeD7qE8EJ1yiLIx4nFGVhLQgmh7uY0nfyhQlJ
-avpXcn+M6HwlaNoFasaLaOknVGh4Rg+DJPGnei0hO9a70LtKZFhqypnYnahJKsF4
-NS3c0WLWwmQ6MfhVPkpjItyqfevtT1jWFUT/QGpcsSOC03cpPrMItcM9AoGALvfj
-uXa8PVDYMBsrUpTjUadONVJ/lPPPdiSnoIqQgoD4PH+M8UcXzScqWxGKvzbhBzOt
-ImnsNwj9okq0KY6+xE4uTEl5Ifa7aK4pB5NTD/UH4khLfr4suuKwx6zi+20TEC7t
-EPtpmotm3tlvNuB62bl9DgB5e3SxQ9JlavPb1m0CgYEAlXchlHeH3Q1xZ/ZBJEU4
-47DuDH/D7NEDrZqp3Vnp++JEmHEF9kCXhC3XEYPfvigBqYHo2sKyUBTw4co9wzDr
-RClkDx5XnZCsXLQkSNcmnJQ+xgmzRKZUR3wOTsDPEHA3LEIS7E8+dlfvYM15Smh4
-+QqKuYGhfc3XT94YGjZaUpc=
------END PRIVATE KEY-----
