--- betterstyles-ts/config/webpack.dev.js
@@ -1,29 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-const webpack = require('webpack');
-const webpackMerge = require('webpack-merge');
-const commonConfig = require('./webpack.common.js');
-
-module.exports = webpackMerge(commonConfig, {
-    devtool: 'eval-source-map',
-    devServer: {
-        publicPath: '/',
-        contentBase: path.resolve('dist'),
-        hot: true,
-        https: {
-            key: fs.readFileSync('./certs/server.key'),
-            cert: fs.readFileSync('./certs/server.crt'),
-            cacert: fs.readFileSync('./certs/ca.crt')
-        },
-        compress: true,
-        overlay: {
-            warnings: false,
-            errors: true
-        },
-        port: 3000,
-        historyApiFallback: true
-    },
-    plugins: [
-        new webpack.HotModuleReplacementPlugin()
-    ]
-});
