--- betterstyles-ts/config/webpack.prod.js
@@ -1,32 +1,0 @@
-const webpack = require('webpack');
-const webpackMerge = require('webpack-merge');
-const commonConfig = require('./webpack.common.js');
-const ENV = process.env.NODE_ENV = process.env.ENV = 'production';
-
-module.exports = webpackMerge(commonConfig, {
-    devtool: 'source-map',
-
-    externals: {
-        'react': 'React',
-        'react-dom': 'ReactDOM'
-    },
-
-    performance: {
-        hints: "warning"
-    },
-
-    plugins: [
-        new webpack.optimize.UglifyJsPlugin({
-            sourceMap: true,
-            mangle: {
-                screw_ie8: true,
-                keep_fnames: true
-            },
-            compress: {
-                warnings: false,
-                screw_ie8: true
-            },
-            comments: false
-        })
-    ]
-});
