--- betterstyles-ts/function-file/function-file.html
@@ -1,22 +1,0 @@
-<!-- Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See full license in root of repo. -->
-<!-- This file shows how to create an empty HTML file to load a function file. -->
-
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=Edge" />
-
-    <!-- Office JavaScript API -->
-    <script type="text/javascript" src="https://appsforoffice.microsoft.com/lib/1.1/hosted/office.debug.js"></script>
-
-    <script type="text/javascript" src="function-file.js"></script>
-</head>
-
-<body>
-  <!-- NOTE: The body is empty on purpose. Since function in function-file.js are
-       invoked via a button, there is no UI to render. -->
-</body>
-
-</html>
