--- betterstyles-ts/function-file/function-file.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
- * See LICENSE in the project root for license information.
- */
-
-(() => {
-  // The initialize function must be run each time a new page is loaded
-  Office.initialize = () => {
-
-  };
-
-  // Add any ui-less function here
-})();
