--- betterstyles-ts/package.json
@@ -1,53 +1,0 @@
-{
-    "name": "betterstyles-ts",
-    "description": "",
-    "author": "",
-    "version": "1.0.0",
-    "scripts": {
-        "clean": "rimraf dist && rimraf .awcache",
-        "lint": "tslint --project tsconfig.json",
-        "start": "webpack-dev-server --inline --config config/webpack.dev.js --progress",
-        "sideload": "office-toolbox sideload -m betterstyles-ts-manifest.xml -a Word",
-        "build": "npm run clean && webpack --config config/webpack.prod.js --colors --progress --bail",
-        "validate": "./node_modules/.bin/validate-office-addin"
-    },
-    "dependencies": {
-        "@material-ui/core": "^3.5.1",
-        "@material-ui/icons": "^3.0.1",
-        "@microsoft/office-js-helpers": "^1.0.1",
-        "babel-loader": "^8.0.4",
-        "core-js": "^2.5.3",
-        "office-ui-fabric-react": "^5.51.0",
-        "react": "^16.2.0",
-        "react-dom": "^16.2.0",
-        "typeface-roboto": "^0.0.54"
-    },
-    "devDependencies": {
-        "@types/office-js": "^0.0.51",
-        "@types/react": "^16.0.38",
-        "@types/react-dom": "^16.0.4",
-        "@types/react-hot-loader": "^3.0.6",
-        "@types/webpack": "2.2.12",
-        "@types/webpack-dev-server": "1.12.6",
-        "autoprefixer": "6.7.7",
-        "copy-webpack-plugin": "4.0.1",
-        "css-loader": "0.27.3",
-        "extract-text-webpack-plugin": "2.1.0",
-        "file-loader": "0.10.1",
-        "html-webpack-plugin": "2.28.0",
-        "less": "^3.0.1",
-        "less-loader": "^4.0.5",
-        "office-addin-validator": "^1.0.1",
-        "office-toolbox": "^0.1.0",
-        "postcss-loader": "1.3.3",
-        "react-hot-loader": "^3.1.3",
-        "rimraf": "2.6.1",
-        "style-loader": "0.16.0",
-        "ts-loader": "^3.5.0",
-        "tslint": "^5.9.1",
-        "typescript": "^2.7.2",
-        "webpack": "^3.11.0",
-        "webpack-dev-server": "^2.11.1",
-        "webpack-merge": "^4.1.1"
-    }
-}
