--- betterstyles-ts/src/index.html
@@ -1,21 +1,0 @@
-<!-- Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. -->
-<!-- See LICENSE in the project root for license information -->
-
-<!doctype html>
-<html lang="en" data-framework="typescript">
-
-<head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=Edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>betterstyles_ts</title>
-
-    <!-- Office JavaScript API -->
-    <script type="text/javascript" src="https://appsforoffice.microsoft.com/lib/1.1/hosted/office.debug.js"></script>
-</head>
-
-<body class="ms-font-m">
-    <div id="container"></div>
-</body>
-
-</html>
