--- betterstyles-ts/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-        "allowJs": true,
-        "target": "es5",
-        "module": "commonjs",
-        "jsx": "react",
-        "moduleResolution": "node",
-        "sourceMap": true,
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "removeComments": false,
-        "outDir": "dist",
-        "allowUnusedLabels": false,
-        "noImplicitReturns": true,
-        "noUnusedParameters": true,
-        "noUnusedLocals": true,
-        "lib": ["es7", "dom"],
-        "pretty": true,
-        "typeRoots": ["node_modules/@types"]
-    },
-    "exclude": ["node_modules"],
-    "compileOnSave": false,
-    "buildOnSave": false
-}
