+++ betterstyles/package.json
@@ -1,34 +1,33 @@
 {
-  "name": "betterstyles",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@material-ui/core": "^3.5.1",
-    "@material-ui/icons": "^3.0.1",
-    "@types/jest": "23.3.9",
-    "@types/material-ui": "^0.21.5",
-    "@types/node": "10.12.6",
-    "@types/react": "16.7.4",
-    "@types/react-dom": "16.0.9",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-scripts": "2.1.1",
-    "typeface-roboto": "^0.0.54",
-    "typescript": "3.2.0-dev.20181025"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+    "name": "betterstyles",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "@material-ui/core": "^3.5.1",
+        "@material-ui/icons": "^3.0.1",
+        "@types/jest": "23.3.9",
+        "@types/node": "10.12.6",
+        "@types/react": "16.7.4",
+        "@types/react-dom": "16.0.9",
+        "react": "^16.6.3",
+        "react-dom": "^16.6.3",
+        "react-scripts": "2.1.1",
+        "typeface-roboto": "^0.0.54",
+        "typescript": "^3.1.6"
+    },
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject"
+    },
+    "eslintConfig": {
+        "extends": "react-app"
+    },
+    "browserslist": [
+        ">0.2%",
+        "not dead",
+        "not ie <= 11",
+        "not op_mini all"
+    ]
 }
